Amendments from "Attack of the NaNs"
Tested the code by bombarding every routine with NaN input (via gsw_check_functions routine, and setting all arguments to NaN). Did not test the validity of the resulting output (afterall, "crap in, crap out ..."), but only three routines gave a `segfault` ...

- gsw_saar
- gsw_deltasa_atlas
- gsw_geo_strf_dyn_height

Fixed by resisting the temptation to specifically test for NaNs (using `isnan`) and instead modified the code to make it more robust to NaN input.

Admittedly this is not an exhaustive test, but hopefully it has picked up the major faults.